test(keypad): drive tests through full parse pipeline
Tests previously injected pre-parsed values, bypassing the parser.
Now exercises parse_advertisement_data end-to-end so byte-layout
regressions in adv_parsers/keypad.py are caught.

- Assert get_battery_percent() instead of raw dict access
- None test exercises parser's own None branch (empty mfr_data)
- Extract make_advertisement_data to tests/__init__.py; remove
  duplicate in test_keypad_vision.py
- Expand SwitchbotKeypad docstring: passive-only, no commands
