+++ switchbot/devices/keypad.py
@@ -6,4 +6,8 @@ from .device import SwitchbotDevice
 
 
 class SwitchbotKeypad(SwitchbotDevice):
-    """Representation of a Switchbot Keypad device."""
+    """Representation of a Switchbot Keypad (WoKeypad) device.
+
+    Passive BLE-only — no commands. Battery and attempt_state come from
+    advertisement parsing in adv_parsers/keypad.py.
+    """
